Add fixups following move of inventory reader to inventory client package
