+++ internal/clients/inventory/testdata/BadYaml/inventory.yaml
